scene: HACK Recalculate positions on scale change
wlr_scene listens to wlr_output_layout's layout_change event, which
is raised when either an output position or its scale is changed.
wlr_scene performs a check and only does something in the event that
output positioning changed, ignoring scale changes.

The scene node/output repositioning code is used to send wp_fractional_scale_v1
preferred_scale hints, and it is therefore important that this triggers
on scale change.

For now, hack it up by disabling the check for repositioning.